+++ tests/testmodule/submodule1.py
@@ -1,0 +1,4 @@
+class Foo2:
+    def __init__(self, arg1, arg2):
+        self.arg1 = arg1
+        self.arg2 = arg2